25 Mar 2025

+++ app/services/expense_service.py
@@ -1,28 +1,38 @@
-from typing import List, Optional
-
-from app.db.database import db
 from app.models.expense_model import Expense
+from app.db.database import db
 
+def create_expense(data): # errors already addressed in controller
+    expense = Expense( # or just do Expense(**data)
+        amount = data["amount"],
+        category = data["category"],
+        date = data["date"]
+    )
+    db.session.add(expense)
+    db.session.commit()
+    return expense
 
-def list_expenses() -> List[Expense]:
-	return Expense.query.order_by(Expense.expense_date.desc()).all()
+def get_all_expenses(): # no error can be created for this task
+    return Expense.query.all()
 
+def get_expense_by_id(id): # the get statement itself returns None
+    return Expense.query.get(id)
 
-def get_expense(expense_id: int) -> Optional[Expense]:
-	return Expense.query.get(expense_id)
+# important to handle errors in functions that change database
+def delete_expense_by_id(id):
+    expense = Expense.query.get(id)
+    if not expense: return None
 
+    db.session.delete(expense)
+    db.session.commit()
+    return True
 
-def create_expense(expense: Expense) -> Expense:
-	db.session.add(expense)
-	db.session.commit()
-	return expense
+def update_expense_by_id(id, data): # check for each col(arg) explicitly
+    expense = Expense.query.get(id)
 
+    if not expense: return None
+    if "amount" in data: expense.amount = float(data["amount"])
+    if "category" in data: expense.category = data["category"]
+    if "date" in data: expense.date = data["date"]
 
-def update_expense(expense: Expense) -> Expense:
-	db.session.commit()
-	return expense
-
-
-def delete_expense(expense: Expense) -> None:
-	db.session.delete(expense)
-	db.session.commit()
+    db.session.commit()
+    return expense
